em: fix float32 forcing in local_em_engine.py and remaining pass-1/pass-2 helpers
Continuation of the double-precision float32-forcing audit (prior commit
b2f23e2b). Fixes local_em_engine.py -- the largest remaining P0 backlog
item, the default exact-local per-iteration engine -- plus more of the
live pass-1/pass-2 code paths:

- local_em_engine.py: noise sigma2/tau2 accumulators now derive from
  precision_policy.score_real_dtype instead of hardcoding float32 (offset/
  sumw scalars, wsum/img_power/norm_correction/a2/xa/scale_xa/scale_aa,
  disabled-noise placeholders, chunked per-block accumulators in both the
  big-JIT-deferred and plain noise branches). _local_mstep_rotations and
  _reorder_bucket_to_indices now preserve source dtype instead of forcing
  float32 (pure select/reorder, no new arithmetic -- same reasoning as the
  already-fixed _decode_dense_best_pose_details). _pad_local_big_jit_
  image_axis derives each padded container's dtype from its source array
  (np.pad already preserves dtype; only the subsequent .astype(np.float32)
  calls were discarding it). The big-JIT arg-prep block (image/scale
  corrections, fourier pre-shifts, translation-sqdist) now uses
  precision_policy.score_real_dtype, matching the correctly-threaded
  normalization_log_z_arg/normalization_log_evidence_arg right next to it.
  _postprocess_local_bucket no longer force-narrows score/posterior/best-
  pose values before writing into the (now correctly-typed) output
  buffers.
- k_class.py: _assemble_result (the universal per-image score/evidence/
  posterior/class-responsibility aggregator for K=1 and K>1) derives its
  output dtype from the per-class stats' own precision instead of
  hardcoding float32 or threading a new parameter through several
  under-instrumented callers. Also fixes make_relion_stats's
  rotation_dtype argument here (that helper's own float32 default was
  silently overriding an otherwise-correct rotation_posterior_sums).
  _sum_noise_stats derives array_dtype from the summed field instead of
  make_noise_stats's hardcoded default. _local_layout_for_class and
  _rotation_prior_with_class_log_prior (K=1/K>1 local-search and sparse-
  pass2 prior paths) fixed the same way.
- significance.py: _compute_k_class_significance_batched (the live pass-1
  entry point) gained a local score_real_dtype derived from its own
  use_float64_scoring param, applied to prior padding, the per-image
  score/evidence/Pmax/class-best-score output buffers, the fused-pass1
  branch, and the final write site.
- preprocessing.py:prepare_batch_preprocess_operands gained a dtype param
  for the per-image scale/normalization correction factors; threaded
  through both live call sites.
- image_shifts.py: integer_pre_shifts_or_none preserves input dtype
  (avoids flipping the integral-shift classification near the atol
  boundary). half_image_phase_factors/tiled_half_image_phase_factors
  gained a dtype param (was hardcoded float32, defeating the
  Precision.HIGHEST matmul inside); threaded through all four live call
  sites.
- projection.py: compute_norm_residual_per_image/
  compute_scale_correction_terms_per_image no longer force-cast to
  float32, matching the fix already applied to their sibling
  compute_noise_block.

Also fixes a real bug this round's own testing caught: removing
compute_norm_residual_per_image's forced cast let float64 values reach
several `array.at[idx].add(value)` JAX scatter-add call sites in
local_em_engine.py/local_big_jit.py/sparse_pass2_bucketed.py, which
(unlike plain +/assignment) require an exact dtype match and were
producing FutureWarnings ("cannot safely cast value from dtype=float64 to
dtype=float32 ... will result in an error"). Fixed at every site by
casting the added value to the destination array's own dtype.

Testing: pixi run test-em-fast-guard 16/16 (before and after the scatter
fix); the two directly-affected tests
(test_run_local_em_exact_default_path_matches_debug_split_path,
..._big_jit_bucket_matches_debug_split) pass cleanly under
-W error::FutureWarning after the fix. Full test_refine_relion_mode.py
(356 tests) run twice against a git-stash-clean baseline of this round's
8 files: 64 failed/291 passed/1 skipped in both runs, byte-identical
FAILED-line sets (confirmed via diff) -- same pre-existing libfftw3.so.3
environment gap as the prior round, zero regressions.

GPU/RELION parity check attempted via scripts/run_multi_iter_parity.py
against the local relion_em_test_double_seeded/ oracle, not completed:
four CPU-only Slurm attempts (up to 300G) all hit OUT_OF_MEMORY at the
same point regardless of allocation size (likely a
psutil.virtual_memory().available-based CPU-mode memory auto-detection
issue independent of this change, worth its own investigation), and a
queued GPU job stayed pending on this account's CPU quota. See
docs/math/relion_parity_agent_notes.md and
docs/math/em_parity_program.md for full detail -- no FSC/RELION-quality
claim is made for these fixes yet.

+++ recovar/em/dense_single_volume/local_big_jit.py
@@ -870,8 +870,13 @@ def run_local_bucket_big_jit(
             ctf2_over_nv_recon = ctf2_over_nv_recon * (batch_scale**2)[:, None]
 
     if apply_fourier_pre_shift:
+        # fourier_pre_shifts already carries precision_policy.score_real_dtype
+        # (set by its caller); pass it through instead of letting these
+        # helpers silently narrow it back to their own float32 default.
         if use_window:
-            pre_shift_phases = half_image_phase_factors(image_shape, fourier_pre_shifts)
+            pre_shift_phases = half_image_phase_factors(
+                image_shape, fourier_pre_shifts, dtype=fourier_pre_shifts.dtype
+            )
             score_phase_expanded = jnp.repeat(pre_shift_phases[:, window_indices], n_trans, axis=0)
             shifted_score = shifted_score * score_phase_expanded
             if not score_only:
@@ -880,7 +885,9 @@ def run_local_bucket_big_jit(
                 if materialize_shifted_noise:
                     shifted_noise = shifted_noise * recon_phase_expanded
         else:
-            phase_expanded = tiled_half_image_phase_factors(image_shape, fourier_pre_shifts, n_trans)
+            phase_expanded = tiled_half_image_phase_factors(
+                image_shape, fourier_pre_shifts, n_trans, dtype=fourier_pre_shifts.dtype
+            )
             shifted_half = shifted_half * phase_expanded
             if not score_only:
                 shifted_recon_half = shifted_recon_half * phase_expanded
@@ -1348,8 +1355,8 @@ def run_local_bucket_big_jit(
             )
             scale_xa_per_image = jnp.where(valid_image_mask, scale_xa_per_image, 0.0)
             scale_aa_per_image = jnp.where(valid_image_mask, scale_aa_per_image, 0.0)
-            noise_scale_xa = noise_scale_xa.at[group_ids].add(scale_xa_per_image)
-            noise_scale_aa = noise_scale_aa.at[group_ids].add(scale_aa_per_image)
+            noise_scale_xa = noise_scale_xa.at[group_ids].add(scale_xa_per_image.astype(noise_scale_xa.dtype))
+            noise_scale_aa = noise_scale_aa.at[group_ids].add(scale_aa_per_image.astype(noise_scale_aa.dtype))
         noise_sigma2_offset = noise_sigma2_offset + noise_sumw_offset
         bucket_norm_correction = batch_img_power_per_image + _compute_norm_residual_per_image(
             proj_for_noise,
